+++ janusgraph-core/src/main/java/org/janusgraph/core/JanusGraph.java
@@ -86,6 +86,11 @@ import org.janusgraph.graphdb.database.cache.CacheInvalidationService;
         test = "org.apache.tinkerpop.gremlin.process.traversal.step.map.MergeVertexTest",
         method = "g_withSideEffectXc_label_person_name_markoX_withSideEffectXm_age_19X_mergeVXselectXcXX_optionXonMatch_selectXmXX_option",
         reason = "JanusGraph doesn't support MergeVertex yet.")
+@Graph.OptOut(
+        test = "org.apache.tinkerpop.gremlin.process.traversal.step.map.CountTest",
+        method = "g_VX1X_valuesXageX_countXlocalX",
+        reason = "TinkerPop 3.6.4 buggy test. We should remove this OptOut when updating to TinkerPop >= 3.6.5 because the test was fixed in TinkerPop 3.6.5. " +
+            "See comment here: https://github.com/JanusGraph/janusgraph/pull/3780#issuecomment-1571961290")
 public interface JanusGraph extends Transaction {
 
     /**